stock predictor

+++ commonsense/optimise/stock.go
@@ -2,6 +2,6 @@ package optimise
 
 import "errors"
 
-func StockPredictor(prices []int) ([]int, error) {
+func StockPredictor(prices []float64) ([]float64, error) {
 	return nil, errors.New("Not implemented")
 }
